% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{depth_result}
\alias{depth_result}
\title{Constructor for depth_result objects}
\usage{
depth_result(depth, max_depth, max_point, max_index)
}
\arguments{
\item{depth}{Numeric vector of depths.}

\item{max_depth}{Maximum depth (scalar).}

\item{max_point}{Coordinates of the point with maximum depth.}

\item{max_index}{Index of the point with maximum depth.}
}
\value{
An object of class \code{depth_result}.
}
\description{
Constructor for depth_result objects
}
